fixed comments of git connection sync module
